fixup! pkg: add Semtech LoRaMAC package
